requirements.txt
# Core dependencies
mcp>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
openai>=1.0.0
anthropic>=0.25.0
numpy>=1.24.0
streamlit>=1.28.0
scipy>=1.11.0
python-dotenv>=1.0.0

# Development dependencies
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
pre-commit>=3.5.0
